+++ headers/convolution.h
@@ -1,1 +1,1 @@
-void linear_convolution(float* x, int N, float* h, int M, double* y, int P);
+void linear_convolution(float* x, int N, float* h, int M, float* y, int P);
